Changes to the structure & create.sh

+++ create.sh
@@ -1,41 +1,82 @@
 #!/bin/bash
 
+while true; do
+    read -p "Do you wish to reset the files on the home folder?" yn
+    case $yn in
+        [Yy]* ) remove=true; break;;
+        [Nn]* ) remove=false; break;;
+        * ) echo "Please answer yes or no.";;
+    esac
+done
+
+function mkdircd () { mkdir -p "$@" && eval cd "\"\$$#\""; }
+
+function remove {
+
+    if [ "$2" = true ]; then
+        sudo rm -f "$1"
+    else
+        rm -f "$1"
+    fi
+
+}
+
 function createLinkIfDoesntExist {
 
+    if [ "$3" = true ]; then
+        remove "$2" "$4"
+    fi
+
 	if [ ! -e $2 ]; then
-		echo "Adding link: $1 -> $2"
-		sudo ln -s "$1" "$2"
+		echo "Adding link: $1 -> $2";
+        if [ "$4" = true ]; then
+            sudo ln -s "$1" "$2";
+        else
+            ln -s "$1" "$2";
+        fi
+
 	else
-		echo "File $2 already exist:"
-		echo "$(ls -la $2)"
+		echo "File $2 already exist:";
+		echo "$(ls -la $2)";
 	fi
 
 }
 
-
 cwd=$PWD
-zshfpath='/usr/share/zsh/site-functions'
+# Copying home files
+cd "$cwd/home"
+for f in *; do
+    if [ "$remove" = true ]; then remove ~/".$f"; fi
+    echo "Adding $cwd/home/$f -> ~/.$f";
+    sed "s|INSTALL_FOLDER|$cwd|g" "$f" >> ~/".$f";
+done
+cd "$cwd"
+
+# Copying Links
+cd "$cwd/links"
+for f in *; do
+    createLinkIfDoesntExist "$cwd/links/$f" ~/".$f" $remove;
+done
+cd "$cwd"
+
 if [ -z "$(which zsh)" ]; then
 	echo "Installing ZSH:"
 	#brew install zsh
 fi
+
+zshfpath='/usr/share/zsh/site-functions'
 if [ -d "$zshfpath" ]; then
-	echo "Adding pureZsh:"
-	createLinkIfDoesntExist "$PWD/zsh/pure/pure.zsh" "$zshfpath/prompt_pure_setup"
+    createLinkIfDoesntExist "$cwd/zsh/pure/pure.zsh" "$zshfpath/prompt_pure_setup" $remove "true";
 fi
-cd ~
-for f in $(ls $cwd/home.*)
-do
-	file=${f/$cwd\/home/}
-	createLinkIfDoesntExist "$f" "$file"
-done
-createLinkIfDoesntExist "$cwd/vim" ".vim"
 
-# Back to normal directory
-cd $cwd
+vundlefolder="$cwd/vim/_bundle/Vundle.vim"
+if [ ! -d "$vundlefolder/.git" ]; then
+    echo "Cloning Vundle repo"
+    mkdircd "$vundlefolder"
+    git clone https://github.com/gmarik/Vundle.vim.git .
+else
+    echo "Vundle Repo already Cloned"
+fi
 
-cd vim	
-mkdir _bundle
-cd _bundle
-git clone https://github.com/gmarik/Vundle.vim.git
-cd ..
+echo "Everything deployed, configure your username/email for git in ~/.giconfig and run :PluginInstall in Vim"
+cd "$cwd"
